Add main and CalculaRegulacao functions
Add main function. Create a percentage vector from 0.3 to 1.5 (30% to 150%) with a difference of 0.001 (0.1%) from
one position to another. Create a electric power vector. Electric power vector is the result of an element-wise
multiplication of the percentage vector and the rated power.

Add CalculaRegulacao function. Add the inputs and outputs and write the formula. Needs to be changed to operate
with vectors and to use the electric power value.